requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.0
redis==5.2.1
rich==13.9.4
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.0
Werkzeug==3.1.3
wrapt==1.17.2
